+++ src/js/config/game.js
@@ -1,301 +1,15 @@
+// gameConfig
 
-import { Keys } from 'eccentric-engine/Engine';
-
+import { phase } from './phase';
 import {
   showScore,
   gameStart,
   spawnWarships,
   nextPhase
-} from './behaviour/gameActions';
-import {
-  PlayerCapsule,
-  PlayerBase
-} from './model/players';
-import {
-  Alien
-} from './model/aliens';
-import { 
-  onPlayerUpdate,
-  playerCapsuleFSMStates, 
-  playerBaseFSMStates, 
-  playerBaseUpdate 
-} from './behaviour/playerActions';
-import {
-  digitFSMStates
-} from './model/score';
-import { 
-  processor as keyProcessor 
-} from './keyProcessor';
-import { 
-  alienUpdate,
-  alienFSMStates 
-} from './behaviour/alienActions';
-import { 
-  PlayerMissile, 
-  AlienBomb 
-} from './model/projectiles';
+} from '../behaviour/gameActions';
+import { processor as keyProcessor } from '../ui/keyProcessor';
 
-const alienTypes = {
-  commandShip: 'commandShip',
-  warship: 'warship',
-  fighter: 'fighter'
-};
-
-const _phases = (phase) => {
-  switch (phase) {
-    case 0: 
-      return {
-        name: 'demo',
-        alienTotal: (alienType) => {
-          return (alienType == 'warship') ? 4 : 3;
-        },
-        alienConcurrent: (alienType) => {
-          return (alienType == 'warship') ? 4 : 3;
-        },
-        alienState: (alienType) => {
-          const states = alienFSMStates();
-          return (alienType == 'warship') ? states.strafe : states.idle;
-        }
-      };
-    case 1:
-      return {
-        name: 'assault',
-        alienTotal: (alienType) => {
-          return (alienType == 'warship') ? 30 : 3;
-        },
-        alienConcurrent: (alienType) => {
-          return (alienType == 'warship') ? 5 : 3;
-        },
-        alienState: (alienType) => {
-          const states = alienFSMStates();
-          return (alienType == 'warship') ? states.hover : states.idle;
-        }
-      };
-    case 2:
-      return {
-        name: 'dive',
-        alienTotal: (alienType) => {
-          return (alienType == 'warship') ? 15 : 0;
-        },
-        alienConcurrent: (alienType) => {
-          return (alienType == 'warship') ? 1 : 0;
-        },
-        alienState: (alienType) => {
-          const states = alienFSMStates();
-          return (alienType == 'warship') ? states.zigZagDive : states.idle;
-        }
-      };
-    case 3:
-      return {
-        name: 'command',
-        alienTotal: (alienType) => {
-          return (alienType == 'commandShip') ? 3 : 0;
-        },
-        alienConcurrent: (alienType) => {
-          return (alienType == 'commandShip') ? 3 : 0;
-        },
-        alienState: (alienType) => {
-          const states = alienFSMStates();
-          return (alienType == 'commandShip') ? states.strafe : states.idle;
-        }
-      };
-    case 4: 
-      return {
-        name: 'bonus',
-        alienTotal: () => {
-          return 0;
-        },
-        alienConcurrent: () => {
-          return 0;
-        }
-      };
-  }
-};
-
-const _digitConfig = () => {
-  return {
-    type: 'digit',
-    width: 55,
-    height: 50,
-    sprites: {
-      digits: {
-        sheet: {
-          path: 'scoresheet.png',
-          frameWidth: 62, 
-          frameHeight: 85,
-          rows: 1,
-          columns: 10
-        },	
-        isDefault: true
-      }
-    },
-    fsmStates: digitFSMStates,
-    update: (digit) => {
-      digit.fsm && digit.fsm.execute();
-    }  
-  }
-};
-
-const _playerCapsuleConfig = () => {
-  return {
-    type: 'playerCapsule',
-    state: 'live',
-    initialVelocity: {
-      x: 0,
-      y: 0
-    },
-    width: 55,
-    height: 50,
-    sprites: {
-      playerCapsule: {
-        sheet: {
-          path: 'player-spritesheet.png',
-          frameWidth: 55, 
-          frameHeight: 50,
-          rows: 1,
-          columns: 5
-        },	
-        isDefault: true
-      }
-    },
-    keys: {
-      left: {
-          keyCodes: [Keys.A, Keys.ARROWLEFT]
-      },
-      right: {
-          keyCodes: [Keys.D, Keys.ARROWRIGHT]
-      },
-      thrust: {
-          keyCodes: [Keys.W, Keys.ARROWUP]
-      },
-      fire: {
-          keyCodes: [Keys.SPACE, Keys.ENTER],
-          minInterval: 1000
-      }
-    },
-    soundEffects: {
-      'shoot': {
-        id: 'playerShoot',
-        type: 'sawtooth',
-        frequency: 2800,
-        volume: 1.0,
-        duration: 80  
-      },
-      'die': {
-        id: 'playerDie',
-        type: 'sawtooth',
-        frequency: 800,
-        volume: 1.0,
-        duration: 120  
-      }
-    },
-    fsmStates: playerCapsuleFSMStates,
-    update: onPlayerUpdate,
-    startRow: 1,
-    startColumn: 2
-  };
-}
-
-const _playerBaseConfig = () => {
-  return {
-    type: 'playerBase',
-    state: 'live',
-    initialVelocity: {
-      x: 0,
-      y: 0
-    },
-    width: 55,
-    height: 50,
-    sprites: {
-      playerBase: {
-        sheet: {
-          path: 'player-base.png',
-          frameWidth: 55, 
-          frameHeight: 50,
-          rows: 1,
-          columns: 1,
-        },	
-        isDefault: true
-      }
-    },
-    fsmStates: playerBaseFSMStates,
-    update: playerBaseUpdate,
-    startRow: 0,
-    startColumn: 2
-  };
-};
-
-const _alienConfig = () => {
-  return {
-    type: undefined,
-    state: undefined,
-    initialVelocity: {
-      x: 0,
-      y: 0
-    },
-    width: 58,
-    height: 50,
-    sprites: {
-      commandShip: {
-        sheet: {
-          path: 'alien-commandship.png',
-          frameWidth: 56, 
-          frameHeight: 50,
-          rows: 1,
-          columns: 1,
-          fromRow: 9,
-          toRow: 9 
-        }  
-      },
-      warship: {
-        sheet: {
-          path: 'alien-hover-spritesheet.png',
-          frameWidth: 58, 
-          frameHeight: 50,
-          rows: 1,
-          columns: 2,
-          fromRow: 7,
-          toRow: 8 
-        }  
-      },
-      fighter: {
-        sheet: {
-          path: 'alien-dive-spritesheet.png',
-          frameWidth: 65, 
-          frameHeight: 60,
-          rows: 1,
-          columns: 5,
-          fromRow: 2,
-          toRow: 6 
-        }  
-      },
-      landedFighter: {
-        sheet: {
-          path: 'alien-landed-spritesheet.png',
-          frameWidth: 58, 
-          frameHeight: 50,
-          rows: 1,
-          columns: 2,
-          fromRow: 1,
-          toRow: 1 
-        }          
-      }
-    },
-    soundEffects: {
-      'die': {
-        id: 'alienDie',
-        type: 'sawtooth',
-        frequency: 3200,
-        volume: 1.0,
-        duration: 80  
-      }
-    },
-    fsmStates: alienFSMStates(),
-    update: alienUpdate
-  }
-};
-
-const _gameConfig = () => {
+export const game = () => {
   return {
     version: 0.1,
     fps: 30,
@@ -350,13 +64,7 @@ const _gameConfig = () => {
         alias: 'viewport'
       }
     },
-    objectTypes: {
-      players: PlayerCapsule.type,
-      aliens: _alienConfig,
-      missiles: PlayerMissile.type,
-      bombs: AlienBomb.type
-    },
-    phases: _phases,
+    phases: phase,
     enableTouchUI: 'auto',
     touchUI: {
       x: ((window.innerWidth / 2) - (300 / 2)),
@@ -530,53 +238,4 @@ const _gameConfig = () => {
       }
     }
   };
-};
-
-class CustomConfig {
-  constructor(customLifecycle) {
-    this._game = _gameConfig();
-    this._game.lifeCycle = {
-      onSetup: customLifecycle.onSetup,
-      onStart: customLifecycle.onStart,
-      onTick: customLifecycle.onTick
-    };
-  };
-  get digit() {
-    return _digitConfig();
-  };
-  get game() {
-    return this._game;
-  };
-  get playerCapsule() {
-    return _playerCapsuleConfig();
-  };
-  get playerBase() {
-    return _playerBaseConfig();
-  };
-  get commandShip() {
-    const conf = _alienConfig();
-    conf.type = alienTypes.commandShip;
-    return conf;
-  };
-  get warship() {
-    const conf = _alienConfig();
-    conf.type = alienTypes.warship;
-    return conf;
-  };
-  get fighter() {
-    const conf = _alienConfig();
-    conf.type = alienTypes.fighter;
-    return conf;
-  };
-  get phases() {
-    return this._game.phases;
-  };
-}
-
-CustomConfig.prototype.phase = function(phaseNumber) {
-  return this.phases[phaseNumber];
-}
-
-export {
-  CustomConfig
 };
